requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
opencv-python>=4.8.0
mss>=9.0.0
numpy>=1.24.0
pillow>=10.0.0